基于多场景训练模型，修改创建配置接口

+++ backend/springbootquickstart/src/main/java/com/easydeploy/springbootquickstart/model/SceneInfo.java
@@ -11,7 +11,7 @@ public class SceneInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long sceneId;
+    private Integer sceneId;
 
     @Column(length = 1000)
     private String sceneDescription;
